+++ Sources/RealityActions/Ease/EaseElasticIn.swift
@@ -10,7 +10,7 @@ import RealityKit
 
 /// Easing function: elastic-in
 public class EaseElasticIn: EaseElastic {
-    public override init (_ action: FiniteTimeAction, period: Float = 0.3) {
+    public override init (_ action: FiniteTimeAction, period: Double = 0.3) {
         super.init(action, period: period)
     }
     
@@ -24,14 +24,14 @@ public class EaseElasticIn: EaseElastic {
 }
 
 class EaseElasticInState: ActionEaseState {
-    let period: Float
+    let period: Double
     
     init? (action: EaseElasticIn, target: Entity) {
         period = action.period
         super.init(action: action, target: target)
     }
     
-    override func update(time: Float) {
+    override func update(time: Double) {
         innerActionState.update (time: elasticIn (time: time, period: period))
     }
 }
